requirements.txt
absl-py==1.3.0
alembic==1.8.1
aniso8601==9.0.1
anyio==3.6.2
apache-airflow==2.4.2
apache-airflow-providers-common-sql==1.2.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-sqlite==3.2.1
apispec==3.3.2
argcomplete==2.0.0
asgiref==3.5.2
asn1crypto==1.5.1
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
Babel==2.10.3
bcrypt==3.2.2
beautifulsoup4==4.11.1
blinker==1.5
Brotli==1.0.9
CacheControl==0.12.11
cachelib==0.9.0
cachetools==5.2.0
cachy==0.3.0
cattrs==22.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cleo==1.0.0a5
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
commonmark==0.9.1
ConfigArgParse==1.5.3
ConfigUpdater==3.1.1
connexion==2.14.1
contourpy==1.0.6
cramjam==2.5.0
crashtest==0.3.1
cron-descriptor==1.2.31
croniter==1.3.7
cryptography==36.0.2
cycler==0.11.0
Deprecated==1.2.13
dill==0.3.6
distlib==0.3.4
Django==4.1.2
djangorestframework==3.14.0
dnspython==2.2.1
docutils==0.19
dulwich==0.20.50
email-validator==1.3.0
et-xmlfile==1.1.0
exceptiongroup==1.0.0
fastapi==0.85.1
fastparquet==0.8.1
filelock==3.7.1
Flask==2.2.2
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-BasicAuth==0.2.0
Flask-Caching==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==22.10.26
fonttools==4.33.3
fsspec==2022.3.0
gast==0.4.0
gevent==22.10.1
geventhttpclient==2.0.8
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.4
graphviz==0.20.1
greenlet==1.1.3.post0
grpcio==1.50.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.7.0
html5lib==1.1
httpcore==0.15.0
httplib2==0.21.0
httpx==0.23.0
idna==3.3
inflection==0.5.1
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
keras==2.10.0
Keras-Preprocessing==1.1.2
keyring==23.9.3
kiwisolver==1.4.2
lazy-object-proxy==1.8.0
libclang==14.0.6
linkify-it-py==2.0.0
lockfile==0.12.2
locust==2.13.0
Mako==1.2.3
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.5.2
mdit-py-plugins==0.3.1
mdurl==0.1.2
mongoengine==0.22.1
more-itertools==9.0.0
msgpack==1.0.4
nltk==3.7
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.0.10
opt-einsum==3.3.0
oscrypto==1.3.0
outcome==1.2.0
packaging==21.3
pandas==1.4.2
paramiko==2.10.4
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
Pillow==9.1.1
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.1.2
portalocker==2.5.1
prison==0.2.1
protobuf==3.19.6
psutil==5.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.14.1
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.0.0
pylev==1.4.0
pymongo==3.12.3
PyMySQL==1.0.2
PyNaCl==1.5.0
pyodbc==4.0.34
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyrsistent==0.19.1
pysftp==0.2.9
PySocks==1.7.1
python-daemon==2.3.2
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
pytz==2022.5
pytzdata==2020.1
PyYAML==6.0
pyzmq==24.0.1
regex==2022.9.13
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.6.0
roundrobin==0.0.4
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
selenium==4.5.0
setproctitle==1.3.2
shellingham==1.5.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
snowflake==0.0.3
snowflake-connector-python==2.7.7
snowflake-sqlalchemy==1.3.4
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.3
starlette==0.20.4
swagger-ui-bundle==0.0.9
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
text-unidecode==1.3
threadpoolctl==3.1.0
tk==0.1.0
tomlkit==0.11.6
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzlocal==1.5.1
uc-micro-py==1.0.1
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.0.1
xattr==0.9.9
zipfile36==0.1.3
zope.event==4.5.0
zope.interface==5.5.0Flask==2.2.2
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-BasicAuth==0.2.0
Flask-Caching==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
